+++ app/src/main/java/rus/one/app/events/Event.kt
@@ -15,7 +15,7 @@ data class Event(
     override val published: String,
     override val coords: Coords?,
     override val link: String?,
-    override val likeOwnerIds: List<Int>?,
+    override val likeOwnerIds: List<Long>?,
     override val likedByMe: Boolean,
     override val attachment: Attachment?,
     override val users: Map<String, User>?,
